+++ lint.sh
@@ -1,3 +1,0 @@
-#!/bin/sh
-
-stack exec hlint -- ./src ./test