+++ changelog.md
@@ -3,6 +3,7 @@
 ## 0.1.0 (2016-09-15)
 * Project setup
 * Card data type
+* Card Bounded and Enum instances
 * newCard function
 * cardRank, cardSuit, cardWord functions
 * Rank and Suit data types