+++ changelog.md
@@ -1,6 +1,6 @@
 # Change Log
 
-## 0.1.0 (2016-09-15)
+## 0.1.0 (2016-09-17)
 * Project setup
 * Card data type
 * Card Bounded and Enum instances